+++ Src/Orbiter/Psys.h
@@ -194,8 +194,8 @@ public:
 
 	void ActivatePlanetLabels(bool activate);
 
-	long FindFirst (int type, _finddata_t *fdata, char *path, char *fname);
-	int FindNext (long fh, _finddata_t *fdata, char *fname);
+	intptr_t FindFirst (int type, _finddata_t *fdata, char *path, char *fname);
+	int FindNext (intptr_t fh, _finddata_t *fdata, char *fname);
 
 private:
 	char *name; // system's name
